requirements.txt
django>5
djangorestframework
mysqlclient
django-cors-headers

